logic to manipulate post
made logic to capture the parts of post that need to have diferent styles.
still dont know if I will use this or if the text and .md that came from github will have a simpler way to transform in html keeping his style, but in anyways it was a good learning.
still missing the corect styling for the post, the style that i know its just to see if its manipulating the text the rigth way.