finished the styles
all the structure are done, now its time to du the functionality. spend a little time on the logic for the manipulation of the text (I improved it in this commit) even if I will probably delete and use a library that will transform the .md in html keeping the format intact.