postList
now the postList reflect the issues. but we still have some problems to solve later.
the onclick is working but I dont think it will work as we want in the future (it will not render the specific post we want)
if we make a search by the searchform the post on the postlist become desconfigured (it only keeps the title) we have to find out why and solve this problem.